Refactor useCalendar hook
Refactor the useCalendar hook into smaller, more manageable files for improved maintainability.
